Port the sales tax calculator spec to Minitest and fixtures
Ports spec/business/sales_tax/sales_tax_calculator_spec.rb to
test/business/sales_tax/sales_tax_calculator_test.rb, keeping all 145
examples and their exact descriptions so the TaxJar cassette lookups keep
resolving. The spec's deep describe/context nesting is preserved through a
small NestedMinitestContext helper that builds real ActiveSupport::TestCase
subclasses, so per-context setup still chains and every example runs once.
Adds the create_bundle_product factory the bundle cases need.

Based on https://github.com/RealBhupesh/gumroad/pull/2
