+++ .env.test
@@ -1,79 +1,1 @@
-DATABASE_HOST="127.0.0.1"
-DATABASE_NAME="gumroad_test"
-TEST_DATABASE_NAME="gumroad_test"
-DATABASE_PORT=3306
-DATABASE_USERNAME="root"
-DATABASE_PASSWORD="password"
-SKIP_NATIVE_MYSQL_RECONNECT="true"
-
-MONGO_DATABASE_URL="localhost:27017"
-MONGO_DATABASE_NAME="gumroad_log_test"
-MONGO_DATABASE_USERNAME=""
-MONGO_DATABASE_PASSWORD=""
-
-ELASTICSEARCH_HOST="http://localhost:9200"
-
-# DBs other than those used in development are specified on purpose so tests don't mess with development data.
-# config/test_redis_isolation.rb rewrites these at boot to a block of databases leased for
-# this run, so concurrent runs cannot flush each other's keys mid-test. These values are the
-# fallback when leasing is skipped (DISABLE_TEST_REDIS_ISOLATION=1, or no free block).
-# That opt-out is read at boot, so under a live spring server it needs DISABLE_SPRING=1
-# or a `bin/spring stop` first.
-REDIS_HOST="localhost:6379/10"
-SIDEKIQ_REDIS_HOST="localhost:6379/11"
-RPUSH_REDIS_HOST="localhost:6379/12"
-RACK_ATTACK_REDIS_HOST="localhost:6379/13"
-
-DEVISE_SECRET_KEY="8cb9397c5059e5d72aac2194f985dfd026423eed76a59692718b1d7bdcc0ea6eb9248a0a7ba91389a470782a4c91e8a9773cfe00737e8edd37bd0a5201bceca1"
-
-OBFUSCATE_IDS_CIPHER_KEY=test-cipher-key
-PAYPAL_PARTNER_MERCHANT_ID=TESTPAYPALID
-PAYPAL_PARTNER_CLIENT_ID=test-paypal-partner-client-id
-PAYPAL_MERCHANT_EMAIL=test-paypal-merchant-email
-GRMC_WEBHOOK_SECRET=test-grmc-webhook-secret
-STRIPE_PUBLIC_KEY_TEST=pk_test_51SNEgsIBOqvOFDrfzBydwy0Ca4j1Xgs7pQsbp2zsyWqHezN7yu3TRignvBqDEc6gn8y98bX8HRv8gKJ8Q45w2t6N00ofXBsMBF
-STRIPE_API_KEY=sk_test_51SNEgsIBOqvOFDrffrs0bKw6Qy174HAkRi4aj9UfzMhPymwCYj6A55u9JahmBn4WJeL8FbawjBmUqUUW6mA1NodU008jShUz2J
-STRIPE__ENDPOINT_SECRET=test-stripe-endpoint-secret
-STRIPE_CONNECT__ENDPOINT_SECRET=test-stripe-connect-endpoint-secret
-
-BRAINTREE_MERCHANT_ID=test-braintree-merchant-id
-BRAINTREE_PUBLIC_KEY=test-braintree-public-key
-BRAINTREE_API_PRIVATE_KEY=test-braintree-private-key
-BRAINTREE_MERCHANT_ACCOUNT_ID_FOR_SUPPLIERS=test-braintree-merchant-account-id
-
-SECURE_ENCRYPT_KEY=1aa983f09f099c052970bc251c1cd5a3
-SECURE_EXTERNAL_ID__PRIMARY_KEY_VERSION=1
-SECURE_EXTERNAL_ID__KEYS__1=a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6
-MOBILE_TOKEN=test-mobile-token
-CLOUDFRONT_PRIVATE_KEY="-----BEGIN RSA PRIVATE KEY-----\nMIIBPAIBAAJBAJ1ooRTziOuIODeLg+aGiLGXP1Qb31LMgzxHmzfgUEVq5RmV8/NY\nYPjfIZ+yDjsZTfP/2H5sgYm4u1PUAiyOqSkCAwEAAQJBAJK+nRU0KNwzRP17Yj8C\nAuNU9mQKC/tbz6jdtLW1t//NvjRRpxZNm9iBIq2CfiWxYTF9anZEjD3Xwv6JIbb4\n8XUCIQDQbE9SP3OY0Y4zjZ0YOEIfKFMc3qtGc9j/3IGbgkGG1wIhAMFXLMxZbFeZ\niYY5yC8MsYCIGEzfraTzJ7W1LMxkG0//AiA4u3Kv8aOWklwBvmdng1DESGavMhEv\nATOtGamR2dfaSwIhAJ+PGDOVYSpyVeLmcOUTbrIKnzNiLeZBnfYCDXSo8Tl1AiEA\nqXZXTJtNlK4SQgnwHXEd+ZmYSGoLb0080MKuh99IOlY=\n-----END RSA PRIVATE KEY-----\n"
-CLOUDFRONT_KEYPAIR_ID=APKAISH5PKOS7WQUJ6SA
-MOBILE_API_OAUTH_APPLICATION_UID=test-mobile-api-oauth-app-uid
-GOOGLE_CLIENT_ID=524830719781-6h0t2d14kpj9j76utctvs3udl0embkpi.apps.googleusercontent.com
-CIRCLE_API_KEY=test-circle-api-key
-DISCORD_BOT_TOKEN=test-discord-bot-token
-
-# MinIO
-AWS_S3_ENDPOINT=http://localhost:9000
-AWS_ACCESS_KEY_ID=minioadmin
-AWS_SECRET_ACCESS_KEY=minioadmin
-S3_DELETER_ACCESS_KEY_ID=minioadmin
-S3_DELETER_SECRET_ACCESS_KEY=minioadmin
-
-AWS_ACCOUNT_ID=123456789012
-
-STRONGBOX_GENERAL="-----BEGIN RSA PRIVATE KEY-----\nProc-Type: 4,ENCRYPTED\nDEK-Info: DES-EDE3-CBC,14957F59FBB1404B\n\ndIBUwXHDIeooeT1u+iBLu9sxbeF0Kh/y2e1fUXhUaRXAVrnEaINrlcAC0OpDxlr8\nZUSZXkukrarCP7qx12B8b6PBzImgOOyy7x5jOVf+iPn0FiUR9HaSg848p2o6tQvt\npdm4FS8EOivBNQBNAK5H1XPuCwBVWEqTrkzQbaR276zX9U3hErtG50Jg1pjQXoqV\n/Agh6LrcSVYmy0otEbviMoEAjW+aA4aWN9KanVG7yoJI0RwQFrXM/v7fDg2lBS4p\nFm4hlXYRoOuLBJ6zm/O0BIyq8TRMgwnAcI1yLxnLkRBPfjROPBboP3UxcpGRejB0\nE5I5odEjvyzQfNGpfDJcESHSUoep/5H2nJ8fMQKHNL7EWnwP/29OmEYQ8hqY7zCL\nkKsWQKBvHfqsYXE9SN8U1i6YU8YOUV1qO2fIXtXrW3T9fJCEL7DEg+vbbzbI/f5R\nH+0m1xDI6qBa+BCpotuizl8hgH4/r3WwqMEP0G9W6w5281JBIcRieg3cOCI7AotQ\nZSHfvWb2ptRBoTi11M3QmlZeq6Zf5eRpVyOXDpBVdFZVaJw63y/yFNb7qJM023I/\naB7KeOSbeFsDz4DF47ViB43MiisJ0CNWkmwow/jMy4EwEjT0EsIf/sXKkPctBjuY\nkdR1JR+Vxe+pipJwzi9aM3zamTyvutBw2rVqzPHM0kb1JIi2wsaiuySReXchUHgM\n9j9DvD2NxegbAjIDdZ+tLt94GVreJSqdPlNItmZh5i6XKsbHDKiujz904qGP36Kh\nfeHLEZBG7ua7SlSkshY9Re7GUmkRNPujAtSbxPwfFpPu30IEGPes1Jyn4Ewlj7g+\n0YzB6f2FXyF1PmUwVfmqnUN2T67jhsr3EL608mV8QBy/HfudJST6lsrgu7VKrtiM\n7AyWW+ryO6EIJH/R117d/2a0QC/kfA7BQ/UQiFa7Bsku3LpQBRRVHTvhUIKKC8ry\nXpH1iecmx1JnIdZXv8u47eieAA+8Kprt5tbyxz09KEuu4jpnFl5Drv6fn7ZPAT8B\nrXFzObvot8OdcjIXnzbcJKp+liHcyOfOTYnit9HnG6xB4X4hEAFWIcu0zO6CfcQa\nuP0yUPnNT71//7nR9AsJ69XeOVjp4gQIfpbqAV31x7Iq3K3dMFPhDCs2v9MpKUYo\n34iZRw3BW9UksxLAWKHUEIS4afAKqz/wCQhXZ4OVJwNF4tZRuVRguLtjuDUiyque\n0ZKmSWi06faMRsYlwnbtZjD2wK4TKxPuystAUQWrImDExvP+pDPU3Y++vAS2ok9Q\nBEYqH9tPdZ7fmD6lFt5xAt88ukT6nWHaNmHjilN1GT/P++z7Z4RIvIk/VQZ6S2qm\n0VytO16EwUEyr3VAEv6k0tZcfOcBiFC7/LuoWlGAFmjBiO7BqQ2VwyhDsZKN+nvq\nx2T5HzjGR+/kXv6vfPWJVUYsDApU7JhsoGPoz4HEPcpEje0PuzkFCDcV798NArn3\nOnH3bgymVFpoUEk5oplTcVNMFrMCn7NheUECzryn4L3Bj20EZYUMdE/+g7HmoXv0\nRpL7N9gHTemG08ufA+0jBqYZoMtcStC5bLP5TqcLaNd6uWCixj3OoYWQiA+fwvE6\n-----END RSA PRIVATE KEY-----\n-----BEGIN PUBLIC KEY-----\nMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAvFtatcBU3q5VKkNpykda\nd48gVT6tQjqNeh07ZYe5Zc1jTjzZ98nSo/KrRfmH0zknR/muc+seBrx3JlwIwrsK\nSisFxaiKkHs/2MpO3FkPS2sfOkwup7tDFOPqQwm12t3KNg7CxP4r1W7YftMYRDzg\nyDntmSprURO9NEMmgQx3ChNlaO1oDbzPDOTFzQR1St7IKhC3nZ4oF5ePwttJHdI9\nsSnjzyuftaSTsk/zm1a4zp2CfSd0L6oWmWX8/3l3fcjn/KcSjZd+hr6kdODO09Ij\na/qx5zUtPWE1W7QfmZFYzNelMv0xuuyGChKpEUJsLZIaqGihbjWAPKwuJEoDQr7w\npwIDAQAB\n-----END PUBLIC KEY-----\n"
-STRONGBOX_GENERAL_PASSWORD="1234"
-
-MAILER_HEADERS_ENCRYPTION_KEY_V1=test_encryption_key
-
-IRAS_API_ID=test-iras-api-id
-IRAS_API_SECRET=test-iras-api-secret
-
-RECAPTCHA_MONEY_SITE_KEY=6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI
-RECAPTCHA_MONEY_SCORE_SITE_KEY=6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_score_key
-RECAPTCHA_LOGIN_SITE_KEY=6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI
-
-INTERNAL_ADMIN_API_TOKEN=test-admin-token
-
-RPUSH_APN_CERT_CREATOR_PASSWORD=mock_creator_password
-RPUSH_APN_CERT_BUYER_PASSWORD=mock_buyer_password
+/Users/gumclaw/repos/gumroad/.env.test
